requirements.txt
pandas~=1.4.2
matplotlib~=3.5.2
opencv-python~=4.8.0.74
numpy~=1.24.4
Pillow~=9.3.0
bigcrittercolor~=0.1.0
scikit-image~=0.19.3
--find-links https://download.pytorch.org/whl/torch_stable.html
torch~=1.11.0+cu113
scikit-learn~=1.1.1
--find-links https://download.pytorch.org/whl/torch_stable.html
torchvision~=0.12.0+cu113
--find-links https://download.pytorch.org/whl/torch_stable.html
torchsampler~=0.1.2
requests~=2.27.1
scipy~=1.8.1
clip~=1.0
ultralytics~=8.0.120
groundingdino~=0.1.0
transformers
addict